backend/apps/alerts/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import SecurityAlertViewSet

router = DefaultRouter()
router.register(r'', SecurityAlertViewSet)

urlpatterns = [path('', include(router.urls))]
